more services done

+++ SystemyWP.API/Data/DTOs/Gastronomy/IngredientDto.cs
@@ -6,14 +6,9 @@ public class IngredientDto
 {
     public long Id { get; set; }
     public string AccessKey { get; set; }
-
-    public string Name { get; set; } 
-
-    public string Description { get; set; } 
-
-    public MeasurementUnits MeasurementUnits { get; set; } 
-
-    public float PricePerStack { get; set; }
-    
-    public float StackSize { get; set; }
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public MeasurementUnits MeasurementUnits { get; set; }
+    public int PricePerStack { get; set; }
+    public double StackSize { get; set; }
 }
